Fixed broken test.
The previous merge of broke one of tests because in the case where there is no Accept header the media-type-available? logic is not called. The merged code had rightly moved complexity out of run-handler into media-type-available? but this case was not accounted for. This has been fixed by adding some special handling the 'false' outcome of the accept-exists? state. This is not ideal, because this special handling would have to be recreated by a developer overriding the accept-exists? step. Potentially we should consider adding in a new step but it would be better to do this once a proper state diagram is published with an accurate map of the liberator states.