enable decisions to return functions as well as truthy values
Small change allows a decision to return a function in the truthy position, thus allowing the execution path to be altered beyond the preset values. Also by exposing defhandler, new handlers can be defined and the execution path can be altered to use these new handlers.

To see more, check out examples/clj/splice.clj which introduces new decision points that an application might use to test the validity of posted or put data and direct to a new unprocessable-entity handler if validation fails.