Improve `combine` to not return lazy sequences
The `combine` function attempts to add new values to a collection by
merging them. However, for lists and a vectors, a lazy sequence is
returned rather than a list or a vector.

For lists, `(apply list (concat ...` is used instead of `(into ...` in
order to preserve element order for backwards compatibility, as `into`
would put the new elements at the head of the list.

For vectors, `into` is used, as new elements are added to the tail.

The implementation for sets is simplified.