Never respond with 304 if last-modified is missing in response
The default modified-since? decision would always evaluate to false
when the request contained an "if-modified-since" header but the
response didn't contain a "last-modified" header. This seems wrong
because the decision is not actually able to determine whether the
resource has been modified or not.

This patch makes it evaluate to true instead which seems more
reasonable in this situation.